Profile dependent SubjectService
